Initial commit - A Python LibGen Interface using the BeautifulSoup web crawler and pandas dataframes. Uses a hard coded search term currently, this will be allowed to be specified in the next release.
Data is parsed using beautifulSoup and is output into a pandas dataframe
for furthur manipulation.

Ideally this will grow into a fledgling API interface, with the hope of
alleviating issues with other methods and interfacing more reliably than
similar projects.

In particular, I had reliability issues with [https://github.com/mmarquezs/libgen-python-api]this project where rows of data would be misaligned with each-other, making it essentially unreliable to work with. Ideally, the source code for this project will be much simpler as well

Further testing is required to ensure that this API is even remotely
robust...